class-work/E222_FINAL/project-code/requirements.txt
markdown
connexion
csv
graphviz
flask
io
IPython.core.interactiveshell
os
request
numpy
matplotlib.pyplot
scipy
sklearn
sklearn.externals.six
